Rename test module to test and add ports as constants
