+++ component/src/test/java/org/ballerinalang/kafka/test/consumer/KafkaConsumerManualCommitTest.java
@@ -16,7 +16,7 @@
 * under the License.
 */
 
-package org.ballerinalang.kafka.nativeimpl.consumer;
+package org.ballerinalang.kafka.test.consumer;
 
 import io.debezium.kafka.KafkaCluster;
 import io.debezium.util.Testing;
@@ -26,12 +26,11 @@ import org.ballerinalang.launcher.util.BCompileUtil;
 import org.ballerinalang.launcher.util.BRunUtil;
 import org.ballerinalang.launcher.util.CompileResult;
 import org.ballerinalang.model.values.BBoolean;
+import org.ballerinalang.model.values.BError;
 import org.ballerinalang.model.values.BInteger;
 import org.ballerinalang.model.values.BMap;
-import org.ballerinalang.model.values.BRefType;
 import org.ballerinalang.model.values.BString;
 import org.ballerinalang.model.values.BValue;
-import org.ballerinalang.model.values.BValueArray;
 import org.ballerinalang.util.codegen.ProgramFile;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
@@ -40,32 +39,34 @@ import org.testng.annotations.Test;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Properties;
 import java.util.concurrent.CountDownLatch;
 
+import static org.ballerinalang.kafka.test.utils.Utils.KAFKA_BROKER_PORT;
+import static org.ballerinalang.kafka.test.utils.Utils.ZOOKEEPER_PORT_1;
 import static org.ballerinalang.kafka.util.KafkaConstants.KAFKA_NATIVE_PACKAGE;
 
 /**
- * Test cases for ballerina.net.kafka consumer ( with seek ) native functions.
+ * Test cases for ballerina.net.kafka consumer ( with manual commit enabled ) manual offset commit
+ * using commit() native function.
  */
 @Test(singleThreaded = true)
-public class KafkaConsumerSeekTest {
+public class KafkaConsumerManualCommitTest {
     private CompileResult result;
     private static File dataDir;
     private static KafkaCluster kafkaCluster;
 
     @BeforeClass
     public void setup() throws IOException {
-        result = BCompileUtil.compile("consumer/kafka_consumer_seek.bal");
+        result = BCompileUtil.compile("consumer/kafka_consumer_manual_commit.bal");
         Properties prop = new Properties();
         kafkaCluster = kafkaCluster().deleteDataPriorToStartup(true)
                 .deleteDataUponShutdown(true).withKafkaConfiguration(prop).addBrokers(1).startup();
         kafkaCluster.createTopic("test", 1, 1);
     }
 
-    @Test(description = "Test Basic consumer with seek")
-    public void testKafkaConsumeWithSeek() {
+    @Test(description = "Test Basic consumer polling with manual offset commit")
+    public void testKafkaConsumeWithManualOffsetCommit() {
         CountDownLatch completion = new CountDownLatch(1);
         kafkaCluster.useTo().produceStrings("test", 10, completion::countDown, () -> "test_string");
         try {
@@ -91,70 +92,18 @@ public class KafkaConsumerSeekTest {
             }
         }
         Assert.assertEquals(msgCount, 10);
+
         ProgramFile programFile = result.getProgFile();
         BMap<String, BValue> part = createPartitionStruct(programFile);
         part.put("topic", new BString("test"));
         part.put("partition", new BInteger(0));
         inputBValues = new BValue[]{consumerEndpoint, part};
 
-
-        returnBValues = BRunUtil.invoke(result, "funcKafkaGetPositionOffset", inputBValues);
-        Assert.assertEquals(returnBValues.length, 1);
+        returnBValues = BRunUtil.invoke(result, "funcKafkaGetCommittedOffset", inputBValues);
         Assert.assertNotNull(returnBValues[0]);
-        //Assert.assertNull(returnBValues[1]);
-        Assert.assertTrue(returnBValues[0] instanceof BInteger);
-        Assert.assertEquals(((BInteger) returnBValues[0]).intValue(), 10);
-
-
-        BMap<String, BValue> offset = createOffsetStruct(programFile);
-        offset.put("partition", part);
-        offset.put("offset", new BInteger(5));
-
-        inputBValues = new BValue[]{consumerEndpoint, offset};
-        BRunUtil.invoke(result, "funcKafkaSeekOffset", inputBValues);
-
-        inputBValues = new BValue[]{consumerEndpoint, part};
-
-        returnBValues = BRunUtil.invoke(result, "funcKafkaGetPositionOffset", inputBValues);
-        Assert.assertEquals(returnBValues.length, 1);
-        Assert.assertNotNull(returnBValues[0]);
-        Assert.assertTrue(returnBValues[0] instanceof BInteger);
-        Assert.assertEquals(((BInteger) returnBValues[0]).intValue(), 5);
-
-        ArrayList<BMap<String, BValue>> structArray = new ArrayList<>();
-        structArray.add(part);
-        BValueArray partitionArray = new BValueArray(structArray.toArray(new BRefType[0]),
-                createPartitionStruct(programFile).getType());
-
-        inputBValues = new BValue[]{consumerEndpoint, partitionArray};
-        returnBValues = BRunUtil.invoke(result, "funcKafkaBeginOffsets", inputBValues);
-        Assert.assertEquals(returnBValues.length, 1);
         Assert.assertTrue(returnBValues[0] instanceof BMap);
-        BMap<String, BValue> off = (BMap<String, BValue>) returnBValues[0];
-        Assert.assertEquals(((BInteger) off.get("offset")).intValue(), 0);
-
-        inputBValues = new BValue[]{consumerEndpoint, partitionArray};
-        returnBValues = BRunUtil.invoke(result, "funcKafkaEndOffsets", inputBValues);
-        Assert.assertEquals(returnBValues.length, 1);
-        Assert.assertTrue(returnBValues[0] instanceof BMap);
-        off = (BMap<String, BValue>) returnBValues[0];
-        Assert.assertEquals(((BInteger) off.get("offset")).intValue(), 10);
-
-        inputBValues = new BValue[]{consumerEndpoint, partitionArray};
-        BRunUtil.invoke(result, "funcKafkaSeekToBegin", inputBValues);
-
-        inputBValues = new BValue[]{consumerEndpoint, part};
-
-        returnBValues = BRunUtil.invoke(result, "funcKafkaGetPositionOffset", inputBValues);
-        Assert.assertEquals(returnBValues.length, 1);
-        Assert.assertNotNull(returnBValues[0]);
-        Assert.assertTrue(returnBValues[0] instanceof BInteger);
-        Assert.assertEquals(((BInteger) returnBValues[0]).intValue(), 0);
-
-        inputBValues = new BValue[]{consumerEndpoint, partitionArray};
-        BRunUtil.invoke(result, "funcKafkaSeekToEnd", inputBValues);
-
-        inputBValues = new BValue[]{consumerEndpoint, part};
+        Assert.assertEquals(((BInteger) ((BMap<String, BValue>) returnBValues[0])
+                .get("offset")).value().intValue(), 0);
 
         returnBValues = BRunUtil.invoke(result, "funcKafkaGetPositionOffset", inputBValues);
         Assert.assertEquals(returnBValues.length, 1);
@@ -163,10 +112,50 @@ public class KafkaConsumerSeekTest {
         Assert.assertEquals(((BInteger) returnBValues[0]).intValue(), 10);
 
         inputBValues = new BValue[]{consumerEndpoint};
+        returnBValues = BRunUtil.invoke(result, "funcKafkaCommit", inputBValues);
+        Assert.assertEquals(returnBValues.length, 1);
+        Assert.assertTrue(returnBValues[0] instanceof BBoolean);
+        Assert.assertTrue(((BBoolean) returnBValues[0]).booleanValue());
+
+
+        inputBValues = new BValue[]{consumerEndpoint, part};
+
+        returnBValues = BRunUtil.invoke(result, "funcKafkaGetCommittedOffset", inputBValues);
+        Assert.assertNotNull(returnBValues[0]);
+        Assert.assertTrue(returnBValues[0] instanceof BMap);
+        Assert.assertEquals(((BInteger) ((BMap<String, BValue>) returnBValues[0])
+                .get("offset")).value().intValue(), 10);
+
+        returnBValues = BRunUtil.invoke(result, "funcKafkaGetPositionOffset", inputBValues);
+        Assert.assertEquals(returnBValues.length, 1);
+        Assert.assertNotNull(returnBValues[0]);
+        Assert.assertTrue(returnBValues[0] instanceof BInteger);
+        Assert.assertEquals(((BInteger) returnBValues[0]).intValue(), 10);
+
+        part.put("topic", new BString("test_not"));
+        part.put("partition", new BInteger(100));
+        inputBValues = new BValue[]{consumerEndpoint, part};
+
+        //test partition which is non existent
+        returnBValues = BRunUtil.invoke(result, "funcKafkaGetCommittedOffset", inputBValues);
+        Assert.assertNotNull(returnBValues[0]);
+
+        returnBValues = BRunUtil.invoke(result, "funcKafkaGetPositionOffset", inputBValues);
+        Assert.assertEquals(returnBValues.length, 1);
+        Assert.assertNotNull(returnBValues[0]);
+        Assert.assertTrue(returnBValues[0] instanceof BError);
+        Assert.assertEquals(((BError) returnBValues[0]).getReason(),
+                "Failed to get position offset: " +
+                        "You can only check the position for partitions assigned to this consumer.");
+
+        part.put("topic", new BString("test"));
+        part.put("partition", new BInteger(0));
+        inputBValues = new BValue[]{consumerEndpoint};
         returnBValues = BRunUtil.invoke(result, "funcKafkaClose", inputBValues);
         Assert.assertEquals(returnBValues.length, 1);
         Assert.assertTrue(returnBValues[0] instanceof BBoolean);
         Assert.assertEquals(((BBoolean) returnBValues[0]).booleanValue(), true);
+
     }
 
     @AfterClass
@@ -186,8 +175,8 @@ public class KafkaConsumerSeekTest {
         if (kafkaCluster != null) {
             throw new IllegalStateException();
         }
-        dataDir = Testing.Files.createTestingDirectory("cluster-kafka-consumer-seek-test");
-        kafkaCluster = new KafkaCluster().usingDirectory(dataDir).withPorts(2181, 9094);
+        dataDir = Testing.Files.createTestingDirectory("cluster-kafka-consumer-manual-commit-test");
+        kafkaCluster = new KafkaCluster().usingDirectory(dataDir).withPorts(ZOOKEEPER_PORT_1, KAFKA_BROKER_PORT);
         return kafkaCluster;
     }
 
@@ -196,11 +185,4 @@ public class KafkaConsumerSeekTest {
                 KAFKA_NATIVE_PACKAGE,
                 KafkaConstants.TOPIC_PARTITION_STRUCT_NAME);
     }
-
-    private BMap<String, BValue> createOffsetStruct(ProgramFile programFile) {
-        return BLangConnectorSPIUtil.createBStruct(programFile,
-                KAFKA_NATIVE_PACKAGE,
-                KafkaConstants.OFFSET_STRUCT_NAME);
-    }
-
 }
