+++ component/src/test/java/org/ballerinalang/kafka/test/services/KafkaConsumerServiceTest.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.ballerinalang.kafka.nativeimpl.services;
+package org.ballerinalang.kafka.test.services;
 
 import io.debezium.kafka.KafkaCluster;
 import io.debezium.util.Testing;
@@ -37,6 +37,8 @@ import java.util.Properties;
 import java.util.concurrent.TimeUnit;
 
 import static org.awaitility.Awaitility.await;
+import static org.ballerinalang.kafka.test.utils.Utils.KAFKA_BROKER_PORT;
+import static org.ballerinalang.kafka.test.utils.Utils.ZOOKEEPER_PORT_1;
 
 /**
  * Test cases for ballerina kafka consumer endpoint bind to a service .
@@ -203,7 +205,7 @@ public class KafkaConsumerServiceTest {
             throw new IllegalStateException();
         }
         dataDir = Testing.Files.createTestingDirectory("cluster-kafka-consumer-service-test");
-        kafkaCluster = new KafkaCluster().usingDirectory(dataDir).withPorts(2183, 9094);
+        kafkaCluster = new KafkaCluster().usingDirectory(dataDir).withPorts(ZOOKEEPER_PORT_1, KAFKA_BROKER_PORT);
         return kafkaCluster;
     }
 }
