+++ component/src/test/java/org/ballerinalang/kafka/test/utils/Utils.java
@@ -1,0 +1,6 @@
+package org.ballerinalang.kafka.test.utils;
+
+public class Utils {
+    public static final int KAFKA_BROKER_PORT = 9094;
+    public static final int ZOOKEEPER_PORT_1 = 2181;
+}
